+++ src/app/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod add;
 pub(crate) mod config_path;
 pub(crate) mod edit;
+pub(crate) mod enable;
 pub(crate) mod init;
 pub(crate) mod list;
 pub(crate) mod r#move;