+++ src/core/mod.rs
@@ -1,5 +1,7 @@
 pub(crate) mod add;
+// pub(crate) mod disable;
 pub(crate) mod edit;
+pub(crate) mod enable;
 pub(crate) mod list;
 pub(crate) mod r#move;
 pub(crate) mod remove;